pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "azure-cognitiveservices-speech>=1.46.0",
    "google-auth>=2.41.1",
    "google-auth-oauthlib>=1.2.2",
    "imageio>=2.37.0",
    "imageio-ffmpeg>=0.6.0",
    "moviepy>=2.2.1",
    "numpy>=2.3.3",
    "openai>=2.0.0",
    "pexels-api>=1.0.1",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "qrcode[pil]>=8.0",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.43",
    "streamlit-oauth>=0.1.14",
    "streamlit>=1.50.0",
    "toml>=0.10.2",
]
